Added documentation and split the code into more appropriate files.
